requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
beautifulsoup4
supabase
python-dotenv
python-multipart
